requirements.txt
pip
django
django-cleanup
django-resized
django-tinymce
djangorestframework
persiantools
gunicorn
django_heroku
PyMySQL
django-mysql
mysqlclient
wheel
dj_database_url
qrcode
openpyxl
pillow
